pyproject.toml
[tool.poetry]
name = "api-gateway-poc"
version = "0.1.0"
description = ""
authors = ["ThaRising <kochbe.ber@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.63.0"
uvicorn = "^0.13.4"
peewee = "^3.14.4"
argon2-cffi = "^20.1.0"
python-jose = {extras = ["cryptography"], version = "^3.2.0"}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
